alterado removeMember para adminAction

+++ back/src/channel/channel.service.ts
@@ -10,7 +10,7 @@ import { UsersRepository } from 'src/users/repositories/users.repository';
 
 import * as bcrypt from 'bcryptjs';
 import { MemberDto } from './dto/member.dto';
-import { RemoveMemberDto } from './dto/removeMember.dto copy';
+import { AdminActionDto } from './dto/adminAction.dto';
 import { LeaveDto } from './dto/leave.dto';
 import { ChannelDto } from './dto/channel.dto';
 import { ChannelType, ChannelMemberStatus } from './constants';
@@ -116,36 +116,45 @@ export class ChannelService {
     }
   }
 
-  async removeMember(memberDto: RemoveMemberDto, userId: any) {
+  async adminAction(adminActionDto: AdminActionDto, userId: any) {
     const isAdmin = await this.repository.checkUser(
-      memberDto.channel_id,
+      adminActionDto.channel_id,
       userId,
     );
     const isOwner = await this.repository.checkOwner(
-      memberDto.channel_id,
+      adminActionDto.channel_id,
       userId,
     );
     const memberIsOwner = await this.repository.checkOwner(
-      memberDto.channel_id,
-      memberDto.member_id,
+      adminActionDto.channel_id,
+      adminActionDto.member_id,
     );
     if ((isAdmin || isOwner) && memberIsOwner == false) {
       const member = await this.repository.checkMember(
-        memberDto.member_id,
-        memberDto.channel_id,
+        adminActionDto.member_id,
+        adminActionDto.channel_id,
       );
       if (!member) {
         throw new NotFoundException('O membro não está no canal');
       }
-      this.repository.removeMemberChannel(
-        memberDto.member_id,
-        memberDto.channel_id,
-      );
-      return 'Membro removido';
+
+      if (adminActionDto.action === 'remove') {
+        this.repository.removeMemberChannel(
+          adminActionDto.member_id,
+          adminActionDto.channel_id,
+        );
+        return 'Membro removido';
+      }
+      if (adminActionDto.action === 'mute') {
+        // colocar função que mute usuário
+        return 'Membro mutado';
+      }
+      if (adminActionDto.action === 'baned') {
+        // colocar função que bane usuário
+        return 'Membro banido';
+      }
     } else {
-      throw new UnauthorizedException(
-        'Você não é admin ou owner, ou não pode excluir o owner',
-      );
+      throw new UnauthorizedException('Você não é admin ou owner');
     }
   }
